Fixed unsigned wraparound problem calculating absolute time from RTP timestamp and RTCP NTP time.
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@33029 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/rtp/rtp_session.cxx
@@ -593,7 +593,7 @@ OpalRTPSession::SendReceiveStatus OpalRTPSession::SyncSource::OnReceiveData(RTP_
 
   PTime absTime(0);
   if (m_syncRealTime.IsValid())
-    absTime = m_syncRealTime + PTimeInterval((frame.GetTimestamp() - m_syncTimestamp)/m_session.m_timeUnits);
+    absTime = m_syncRealTime + PTimeInterval(((int64_t)frame.GetTimestamp() - (int64_t)m_syncTimestamp) / m_session.m_timeUnits);
   frame.SetAbsoluteTime(absTime);
 
 #if OPAL_RTCP_XR
